Move config and db fns out of DMM.
